refactor: async config 모듈 변경
